+++ src/main/java/com/github/bluekey/config/security/jwt/JwtProvider.java
@@ -1,4 +1,4 @@
-package com.github.bluekey.jwt;
+package com.github.bluekey.config.security.jwt;
 
 import com.github.bluekey.entity.member.MemberRole;
 import com.github.bluekey.entity.member.MemberType;
